+++ src/consumer/consumer.js
@@ -39,8 +39,10 @@ import MissingFilesFromComponent from './component/exceptions/missing-files-from
 import ComponentNotFoundInPath from './component/exceptions/component-not-found-in-path';
 import npmClient from '../npm-client';
 import { RemovedLocalObjects } from '../scope/component-remove';
-import { linkComponents, linkAllToNodeModules } from '../links';
+import { linkComponents, linkAllToNodeModules, linkComponentsToNodeModules } from '../links';
 import * as packageJson from './component/package-json';
+import Remotes from '../remotes/remotes';
+import type { PathChangeResult } from './bit-map/bit-map';
 
 export type ConsumerProps = {
   projectPath: string,
@@ -503,13 +505,12 @@ export default class Consumer {
     excludeRegistryPrefix?: boolean
   }): Promise<Component[]> {
     const dependenciesIdsCache = [];
-    const remotes = await this.scope.remotes();
+    const remotes: Remotes = await this.scope.remotes();
     const writeComponentsP = componentsWithDependencies.map((componentWithDeps: ComponentWithDependencies) => {
       const bitDir = writeToPath || this.composeComponentPath(componentWithDeps.component.id);
-      componentWithDeps.component.writtenPath = bitDir;
-      // if a component.scope is listed as a remote, it doesn't go to the hub and therefore it can't import dependencies as packages
+      // if a it doesn't go to the hub, it can't import dependencies as packages
       componentWithDeps.component.dependenciesSavedAsComponents =
-        saveDependenciesAsComponents || remotes.get(componentWithDeps.component.scope);
+        saveDependenciesAsComponents || !remotes.isHub(componentWithDeps.component.scope);
       // AUTHORED and IMPORTED components can't be saved with multiple versions, so we can ignore the version to
       // find the component in bit.map
       const componentMap = this.bitMap.getComponent(componentWithDeps.component.id.toStringWithoutVersion(), false);
@@ -599,7 +600,7 @@ export default class Consumer {
 
     await this.bitMap.write();
     if (installNpmPackages) await this.installNpmPackagesForComponents(componentsWithDependencies, verbose);
-    if (addToRootPackageJson) await packageJson.addComponentsToRoot(this, writtenComponents);
+    if (addToRootPackageJson) await packageJson.addComponentsToRoot(this, writtenComponents.map(c => c.id));
 
     return linkComponents(componentsWithDependencies, writtenComponents, writtenDependencies, this, createNpmLinkFiles);
   }
@@ -607,9 +608,9 @@ export default class Consumer {
   moveExistingComponent(component: Component, oldPath: string, newPath: string) {
     if (fs.existsSync(newPath)) {
       throw new Error(
-        `could not move the component ${
-          component.id
-        } from ${oldPath} to ${newPath} as the destination path already exists`
+        `could not move the component ${component.id} from ${oldPath} to ${
+          newPath
+        } as the destination path already exists`
       );
     }
     const componentMap = this.bitMap.getComponent(component.id);
@@ -668,16 +669,17 @@ export default class Consumer {
 
       version.log = componentFromModel.log; // ignore the log, it's irrelevant for the comparison
 
-      // sometime dependencies from the FS don't have an exact version, copy the version from the model
+      // sometime dependencies from the FS don't have an exact version.
+      // in case of auto-tag, the files can be identical between the model and the FS, but the dependency version would
+      // be different.
+      // we don't want to show the component as modified in such cases, so copy the version from the model
       version.dependencies.forEach((dependency) => {
-        if (!dependency.id.hasVersion()) {
-          const idWithoutVersion = dependency.id.toStringWithoutVersion();
-          const dependencyFromModel = componentFromModel.dependencies.find(
-            modelDependency => modelDependency.id.toStringWithoutVersion() === idWithoutVersion
-          );
-          if (dependencyFromModel) {
-            dependency.id = dependencyFromModel.id;
-          }
+        const idWithoutVersion = dependency.id.toStringWithoutVersion();
+        const dependencyFromModel = componentFromModel.dependencies.find(
+          modelDependency => modelDependency.id.toStringWithoutVersion() === idWithoutVersion
+        );
+        if (dependencyFromModel) {
+          dependency.id = dependencyFromModel.id;
         }
       });
 
@@ -840,7 +842,7 @@ export default class Consumer {
     return path.join(this.getPath(), dependenciesDir, bitId.toFullPath());
   }
 
-  async movePaths({ from, to }: { from: string, to: string }) {
+  async movePaths({ from, to }: { from: string, to: string }): PathChangeResult[] {
     const fromExists = fs.existsSync(from);
     const toExists = fs.existsSync(to);
     if (fromExists && toExists) {
@@ -856,6 +858,12 @@ export default class Consumer {
       fs.moveSync(from, to);
     }
     await this.bitMap.write();
+    if (!R.isEmpty(changes)) {
+      const componentsIds = changes.map(c => BitId.parse(c.id));
+      await packageJson.addComponentsToRoot(this, componentsIds);
+      const { components } = await this.loadComponents(componentsIds);
+      linkComponentsToNodeModules(components, this);
+    }
     return changes;
   }
 
@@ -1175,5 +1183,13 @@ export default class Consumer {
       .filter(dir => dir);
     await this._installPackages(dirs, verbose, true);
     return linkAllToNodeModules(this);
+  }
+
+  async eject(componentsIds) {
+    const componentIdsWithoutScope = componentsIds.map(id => id.toStringWithoutScope());
+    await this.remove(componentIdsWithoutScope, true, false);
+    await packageJson.addComponentsWithVersionToRoot(this, componentsIds);
+    await this._installPackages([], true, true);
+    return componentsIds;
   }
 }